Added PersonalMemories Module via upload
Module to be used with the MagicMirror open source program. It displays images/videos whose file names are extracted for an sqlite3 database. It displays the images/videos in a rotating carousel style. It displays 5 images/videos at a time. For my project I am using it to display memories(video/images) submitted by my friend's friends and family. Captions are displayed above the memory; if any.